requirements.txt
s5cmd
polars
pandas
pyarrow
pycolmap
matplotlib
plotly
nbformat
scikit-learn
scipy
ipywidgets
pillow
nbconvert